Improve Geolocation ports
