+++ src/Apps/LocationPicker/Messages.elm
@@ -1,8 +1,9 @@
 module Apps.LocationPicker.Messages exposing (Msg(..))
 
 import Json.Encode exposing (Value)
+import Utils.Ports.Geolocation as Geolocation
 
 
 type Msg
     = MapClick Value
-    | GeoResp Value
+    | GeolocationMsg Geolocation.Id Geolocation.Msg
