+++ src/Apps/LocationPicker/Subscriptions.elm
@@ -1,7 +1,7 @@
 module Apps.LocationPicker.Subscriptions exposing (..)
 
 import Utils.Ports.Map as Map
-import Utils.Ports.Geolocation exposing (geoLocResp)
+import Utils.Ports.Geolocation as Geolocation
 import Apps.LocationPicker.Config exposing (..)
 import Apps.LocationPicker.Models exposing (Model)
 import Apps.LocationPicker.Messages exposing (Msg(..))
@@ -11,5 +11,5 @@ subscriptions : Config msg -> Model -> Sub msg
 subscriptions config model =
     Sub.batch
         [ Map.mapClick (MapClick >> config.toMsg)
-        , geoLocResp (GeoResp >> config.toMsg)
+        , Sub.map config.toMsg <| Geolocation.subscribe GeolocationMsg
         ]
