+++ src/Apps/LocationPicker/Update.elm
@@ -3,7 +3,7 @@ module Apps.LocationPicker.Update exposing (update)
 import Utils.React as React exposing (React)
 import Json.Decode exposing (Value)
 import Utils.Ports.Map as Map
-import Utils.Ports.Geolocation as Gloc
+import Utils.Ports.Geolocation as Geolocation
 import Apps.LocationPicker.Config exposing (..)
 import Apps.LocationPicker.Models exposing (..)
 import Apps.LocationPicker.Messages as LocationPicker exposing (Msg(..))
@@ -24,8 +24,11 @@ update config msg model =
         MapClick value ->
             onMapClick value model
 
-        GeoResp value ->
-            onGeoResp config value model
+        GeolocationMsg id msg ->
+            if id == model.self then
+                onGeoMsg config msg model
+            else
+                ( model, React.none )
 
 
 onMapClick : Value -> Model -> UpdateResponse msg
@@ -40,33 +43,14 @@ onMapClick value model =
         ( model_, React.none )
 
 
-onGeoResp : Config msg -> Value -> Model -> UpdateResponse msg
-onGeoResp config value model =
-    if Gloc.checkInstance value model.self then
-        geoResp config value model
-    else
-        ( model, React.none )
+onGeoMsg : Config msg -> Geolocation.Msg -> Model -> UpdateResponse msg
+onGeoMsg config msg model =
+    case msg of
+        Geolocation.Coordinates lat lng ->
+            ( model.mapEId, lat, lng, 18 )
+                |> Map.mapCenter
+                |> React.cmd
+                |> (,) model
 
-
-geoResp : Config msg -> Value -> Model -> UpdateResponse msg
-geoResp config value model =
-    let
-        newPos =
-            value
-                |> Map.decodeCoordinates
-                |> Result.toMaybe
-
-        model_ =
-            setPos newPos model
-
-        react =
-            case newPos of
-                Just { lat, lng } ->
-                    [ Map.mapCenter ( model.mapEId, lat, lng, 18 ) ]
-                        |> List.map React.cmd
-                        |> React.batch config.batchMsg
-
-                Nothing ->
-                    React.none
-    in
-        ( model_, react )
+        _ ->
+            ( model, React.none )
