+++ src/Setup/Pages/PickLocation/Messages.elm
@@ -1,11 +1,11 @@
 module Setup.Pages.PickLocation.Messages exposing (..)
 
 import Json.Encode exposing (Value)
+import Utils.Ports.Geolocation as Geolocation
 
 
 type Msg
     = MapClick Value
-    | GeoLocResp Value
-    | GeoRevResp Value
+    | GeolocationMsg Geolocation.Id Geolocation.Msg
     | ResetLoc
     | Checked (Maybe String)
