+++ src/Setup/Pages/PickLocation/Subscriptions.elm
@@ -1,7 +1,7 @@
 module Setup.Pages.PickLocation.Subscriptions exposing (subscriptions)
 
 import Utils.Ports.Map exposing (mapClick)
-import Utils.Ports.Geolocation exposing (geoLocResp, geoRevResp)
+import Utils.Ports.Geolocation as Geolocation
 import Setup.Pages.PickLocation.Models exposing (..)
 import Setup.Pages.PickLocation.Messages exposing (..)
 import Setup.Pages.PickLocation.Config exposing (..)
@@ -11,6 +11,5 @@ subscriptions : Config msg -> Model -> Sub msg
 subscriptions { toMsg } model =
     Sub.batch
         [ mapClick <| MapClick >> toMsg
-        , geoLocResp <| GeoLocResp >> toMsg
-        , geoRevResp <| GeoRevResp >> toMsg
+        , Sub.map toMsg <| Geolocation.subscribe GeolocationMsg
         ]
