+++ static/js/index.js
@@ -9,4 +9,4 @@ require("leaflet_js");
 require('./onload.js');
 require('./audio.js');
 require('./map.js');
-require('./geoloc.js');
+require('./geolocation.js');
